Made prettier terrain
Added some ground clutter to the terrain, to make it more "live". Eventual collision issues are *not* accounted for, as I'm sure someone else is better at handling that.

Colors might be somewhat 'off', but that can be fixed at a moments notice.

The new stage isn't added to the demo scene, but is made as a prefab to be added.