Added timer
It should display at the beginning of the Setup Phase, and follow the countdown.

Once the lower seconds are left (and sounds play for each), the color should change to yellow and finally red, while the size increases each time.

Once the Battle Phase starts, it is set to be unactive, and disappears. Currently, there's no consideration for any reset, but can easily be added later.