Added player interaction, to heal the characters upon double-tapping them.
The main functionality lies in the CombatTouchPanels, under Menu, while delegate functionality lies on the individual touch panels below it.

On the main functionalities, you can tweak the healing effect, and the amount of heals each player is given each round. Heals used are reset between rounds.

Sizes of touch panels may need to be adjusted on the SUR40.

Tested at home, while disabling SpawnFromCards(), but it should be restored at this point and not pose a problem.